pyproject.toml
[tool.poetry]
name = "square_image_cropper"
version = "0.1.0"
description = ""
authors = ["meg-ssk"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
opencv-python = "^4.6.0"
PyYAML = "^6.0"
tqdm = "^4.64.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
